feat: update purchase
