+++ dapps/payment/src/modules/Consumer/PaidAPPs/index.tsx
@@ -12,6 +12,7 @@ import Withdraw from 'payment/src/modules/Common/Withdraw';
 import { useBoundProviderStore } from 'payment/src/store';
 import { PAYMENT_TYPE } from 'payment/src/utils/constants';
 import BigNumber from 'bignumber.js';
+// import PurchaseSubscription from 'payment/src/modules/Common/PurchaseSubscription';
 
 const { Search } = Input;
 
@@ -43,6 +44,10 @@ export default () => {
         account && fetch(account);
     }, [account]);
 
+    // const handleComplate = useCallback(() => {
+    //     account && fetch(account);
+    // }, []);
+
     const columns = useMemo(
         () =>
             [
@@ -64,8 +69,6 @@ export default () => {
                                 new BigNumber(row.billing.deferTimeSecs).plus(row.billing.withdrawSchedule).lt(+new Date() / 1000);
                             const isFrozen = row.billing.deferTimeSecs !== '0' || isWithdrawable;
                             const isRefundable = row.billing.balance !== '0' && !isWithdrawable;
-
-                            // console.log('row: ', row);
 
                             return (
                                 <div className="flex align-middle flex-wrap">
@@ -106,6 +109,7 @@ export default () => {
                                             Details
                                         </Link>
                                     </Button>
+                                    {/* <PurchaseSubscription appAddr={row.address} onComplete={handleComplate} /> */}
                                 </div>
                             );
                         }
