+++ dapps/payment/src/utils/request.ts
@@ -18,7 +18,6 @@ import { CONTRACT_ABI } from 'payment/src/contracts/constants';
 import { personalSign } from '@cfxjs/use-wallet-react/ethereum';
 // @ts-ignore
 import { binary_to_base58 } from 'base58-js';
-import Subscription from 'js-conflux-sdk/dist/types/subscribe/Subscription';
 import { ONE_DAY_SECONDS } from './constants';
 
 interface RequestProps {
@@ -35,13 +34,11 @@ interface ErrorType {
 type EditableAPI = Pick<ResourceDataSourceType, 'resourceId' | 'index' | 'op' | 'weight'>;
 
 const INTERFACE_APPV2 = new ethers.utils.Interface(CONTRACT_ABI['appv2']);
-const INTERFACE_APPCoin = new ethers.utils.Interface(CONTRACT_ABI['appCoin']);
 const INTERFACE_VIPCoin = new ethers.utils.Interface(CONTRACT_ABI['vipCoin']);
 const CONTRACT_APPREGISTRY = getContract('appRegistry');
 
 const INTERFACE_APP = new ethers.utils.Interface(CONTRACT_ABI['app']);
 const MULTICALL = getContract('multicall');
-const CONTRACT_CONTROLLER = getContract('controller');
 
 const noticeError = (e: unknown) => {
     let msg = '';
@@ -176,7 +173,7 @@ export const getPaidAPPs = async (account: string) => {
                 },
                 subscription: {
                     name: a.vipCardName,
-                    expired: a.vipExpireAt.toNumber(),
+                    expired: Number(a.vipExpireAt.toString()),
                 },
             })),
             total: r.total,
@@ -188,7 +185,7 @@ export const getPaidAPPs = async (account: string) => {
     }
 };
 
-export const postAPP = async ({ name, url, weight, description, symbol, account, type }: PostAPPType) => {
+export const postAPP = async ({ name, url, weight, description = '', symbol, account, type }: PostAPPType) => {
     try {
         await (
             await CONTRACT_APPREGISTRY.connect(signer).create(
@@ -235,57 +232,6 @@ export const getAPP = async (address: string): Promise<APPDetailType> => {
             description: '',
         };
     }
-    // try {
-    //     const calls: Array<[string, any[]?]> = [
-    //         ['name'],
-    //         ['symbol'],
-    //         ['appOwner'],
-    //         ['totalCharged'],
-    //         ['totalRequests'],
-    //         ['listUser', [0, 0]],
-    //         ['listResources', [0, 0]],
-    //         ['totalTakenProfit'],
-    //     ];
-    //     if (account) {
-    //         calls.push(['frozenMap', [account]]);
-    //     }
-    //     const promises = calls.map((c) => [address, INTERFACE_APP.encodeFunctionData(...c)]);
-    //     const results: { returnData: ethers.utils.Result } = await MULTICALL.callStatic.aggregate(promises);
-    //     const r = results.returnData.map((d, i) => INTERFACE_APP.decodeFunctionResult(calls[i][0], d));
-
-    //     return {
-    //         name: r[0][0],
-    //         link: r[1][0],
-    //         owner: r[2][0],
-    // earnings: formatNumber(r[3][0].sub(r[7][0]), {
-    //     limit: 0,
-    //     decimal: 18,
-    // }),
-    //         requests: r[4][0].toNumber(),
-    //         users: r[5]['total'].toNumber(),
-    //         resources: {
-    //             list: [],
-    //             total: r[6][1].toNumber(),
-    //         },
-    //         frozen: account ? r[calls.length - 1][0].toString() : '0',
-    //     };
-    // } catch (error) {
-    //     console.log('getAPP error: ', error);
-    //     noticeError(error);
-    //     return {
-    //         name: '',
-    //         link: '',
-    //         owner: '',
-    //         earnings: 0,
-    //         requests: 0,
-    //         users: 0,
-    //         resources: {
-    //             list: [],
-    //             total: 0,
-    //         },
-    //         frozen: '0',
-    //     };
-    // }
 };
 
 export const getAPPAPIs = async (address: RequestProps['address']): Promise<APPResourceType> => {
@@ -471,77 +417,6 @@ export const deposit = async ({ amount, appAddr }: { amount: string; appAddr: st
     }
 };
 
-export const getPaidAPPs_bak = async (account: string) => {
-    try {
-        const apps = await CONTRACT_APPREGISTRY.listByUser(account, 0, 1e8);
-        const appContracts = apps[1];
-        const APPsDetail = await getAPPsDetail(appContracts.map((app: any) => app.addr));
-
-        console.log('APPsDetail: ', APPsDetail);
-
-        return [];
-
-        // const Interface = new ethers.utils.Interface(CONTRACT_ABI['app']);
-        // const apiAddr = await CONTRACT_CONTROLLER.api();
-        // const apiContract = getContract('api', apiAddr);
-        // const apps = await apiContract.listPaidApp(account, 0, 1e15);
-        // const appContracts: string[] = apps[0];
-
-        // const calls: ContractCall[] = [
-        //     ['name'],
-        //     ['symbol'],
-        //     ['appOwner'],
-        //     ['totalCharged'],
-        //     ['balanceOfWithAirdrop', [account]],
-        //     ['frozenMap', [account]],
-        //     ['forceWithdrawDelay'],
-        // ];
-
-        // const promises = lodash.flattenDepth(
-        //     appContracts.map((a) => calls.map((c) => [a, Interface.encodeFunctionData(...c)])),
-        //     1
-        // );
-
-        // const results: { returnData: ethers.utils.Result } = await MULTICALL.callStatic.aggregate(promises);
-
-        // const data = lodash
-        //     .chunk(results.returnData, calls.length)
-        //     .map((r) => {
-        //         return r.map((d, i) => {
-        //             return Interface.decodeFunctionResult(calls[i][0], d);
-        //         });
-        //     })
-        //     .map((d, i) => {
-        //         return {
-        //             address: appContracts[i],
-        //             name: d[0][0],
-        //             link: d[1][0],
-        //             owner: d[2][0],
-        //             earnings: formatNumber(d[3][0], {
-        //                 limit: 0,
-        //                 decimal: 18,
-        //             }),
-        //             balance: formatNumber(d[4].total.sub(d[4].airdrop_), {
-        //                 limit: 0,
-        //                 decimal: 18,
-        //             }),
-        //             airdrop: formatNumber(d[4].airdrop_, {
-        //                 limit: 0,
-        //                 decimal: 18,
-        //             }),
-        //             frozen: d[5][0].toString(),
-        //             forceWithdrawDelay: d[6][0].toString(),
-        //         };
-        //     });
-
-        // return data;
-    } catch (error) {
-        console.log('getPaidAPPs error: ', error);
-        noticeError(error);
-        return [];
-    }
-};
-
 export const getAPIKey = async (appAddr: string) => {
     try {
         const msg = { domain: 'web3pay', contract: appAddr };
@@ -639,7 +514,7 @@ export const configAPPCard = async (address: RequestProps['address'], data: any)
     }
 };
 
-export const purchaseCard = async (appAddr: RequestProps['address'], templateId: string, amount: number | string) => {
+export const purchaseSubscription = async (appAddr: RequestProps['address'], templateId: string, amount: number | string) => {
     try {
         const contracts = await getAPPsRelatedContract([appAddr].map((app: any) => app));
         const r = await (
@@ -650,7 +525,7 @@ export const purchaseCard = async (appAddr: RequestProps['address'], templateId:
         return r;
     } catch (error) {
         noticeError(error);
-        return {};
+        throw error;
     }
 };
